added googlePathImage to it's own component, renamed utils to dataFetcher for better clarity
